+++ ving/jobs/handlers/DeleteUnusedS3File.mjs
@@ -1,4 +1,5 @@
 import ving from '#ving/index.mjs';
+import { useKind } from '#ving/record/utils.mjs';
 
 /**
  * This handler deletes a S3File if it does not achieve a ready state by the time this job runs. 
@@ -7,7 +8,7 @@ import ving from '#ving/index.mjs';
  */
 export default async function (job) {
     ving.log('jobs').info(`Instanciating S3File ${job.data.id}`);
-    const s3files = await ving.useKind('S3File');
+    const s3files = await useKind('S3File');
     const s3file = await s3files.findOrDie(job.data.id);
     if (s3file.get('status') == 'ready') {
         ving.log('jobs').info(`S3File ${job.data.id} is in ready state.`);